move common types to types.ts
