+++ packages/react-async/src/reducer.ts
@@ -8,7 +8,7 @@ import {
   AsyncRejected,
   AsyncInitial,
   AbstractState,
-} from "./Async"
+} from "./types"
 
 /**
  * Due to https://github.com/microsoft/web-build-tools/issues/1050, we need
